ประเมินชุมชน

+++ app/Http/Controllers/EvaluateController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Exception;
 use App\Models\Part;
+use App\Models\User;
+use App\Models\PartType;
 use App\Models\PartTarget;
 use Illuminate\Http\Request;
 use App\Models\PartTargetSub;
@@ -13,10 +15,9 @@ use App\Models\AppraisalQuestion;
 use App\Models\PartIndexQuestion;
 use Illuminate\Support\Facades\DB;
 use App\Models\AppraisalTransaction;
-use App\Models\PartType;
-use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
@@ -32,9 +33,28 @@ class EvaluateController extends Controller
         $part_type = PartType::all();
         $part = Part::all();
 
+        $link = DB::select("select config_value from config where config_key = 'community-all'");
+        $api_community = Http::get($link[0]->config_value);
+        $response_community = $api_community->json('data');
+        $session_api_community_name = session()->get('community_name');
+        $session_api_community_id = "";
+
+        if($session_api_community_name != "")
+        {
+            for ($i=0; $i < count($response_community); $i++) { 
+                if($session_api_community_name == $response_community[$i]['community_name'])
+                {
+                    $session_api_community_id = $response_community[$i]['community_id'];
+                    session()->put('session_api_community_id', $session_api_community_id);
+                }
+            }
+        }
+        
         return view('evaluate.index', [
             'part' => $part,
             'part_type' => $part_type,
+            'response_community' => $response_community,
+            'session_api_community_id' => $session_api_community_id,
         ]);
     }
 
@@ -53,19 +73,15 @@ class EvaluateController extends Controller
     public function target($part_id) //เกณฑ์การประเมิน 
     {
         $part = Part::find($part_id);
+        $part_type_id = $part->part_type_id;
+        $part_type = PartType::where('part_type_id', $part_type_id)->get();
+        $part_type_name = $part_type[0]->part_type_name;
         $part_target = PartTarget::where('part_id', $part_id)->get();
-        //สถานะ
-        $transaction = AppraisalTransaction::all();
-        //คะแนน
-        $score = AppraisalScore::all();
-        $part_target_sub = PartTargetSub::all();
-
+        
         return view('evaluate.target', [
             'part' => $part,
             'part_target' => $part_target,
-            'transaction' => $transaction,
-            'score' => $score,
-            'part_target_sub' => $part_target_sub,
+            'part_type_name' => $part_type_name,
         ]);
     }
 
@@ -89,13 +105,14 @@ class EvaluateController extends Controller
         $part_index_question = PartIndexQuestion::orderBy('part_index_question_order', 'asc')->get();
 
         //status
-        $transaction = AppraisalTransaction::select('appraisal_transaction_status')->where('part_target_id', $part_target_id)->get();
+        $community_name = session()->get('community_name');
+        $transaction = AppraisalTransaction::select('appraisal_transaction_status')->where(['part_target_id' => $part_target_id, 'community_name' => $community_name])->get();
 
         if (!empty($transaction[0]->appraisal_transaction_status)) {
             if ($transaction[0]->appraisal_transaction_status == '1') {
                 //เรียกข้อมูลแบบร่าง
-                $ap_question = AppraisalQuestion::select('*')->where('part_target_id', $part_target_id)->get();
-                $ap_score = AppraisalScore::select('*')->where('part_target_id', $part_target_id)->get();
+                $ap_question = AppraisalQuestion::select('*')->where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->get();
+                $ap_score = AppraisalScore::select('*')->where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->get();
                 // dd($ap_score);
                 return view('evaluate.form-draft', [
                     'part' => $part,
@@ -135,19 +152,19 @@ class EvaluateController extends Controller
         WHERE part_target_id = $part_target_id
         ");
 
-        $countQuestion = AppraisalQuestion::where('part_target_id', $part_target_id)->count();
+        $countQuestion = AppraisalQuestion::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->count();
         if ($countQuestion > 0) {
-            AppraisalQuestion::where('part_target_id', $part_target_id)->delete();
+            AppraisalQuestion::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->delete();
         }
 
-        $countScore = AppraisalScore::where('part_target_id', $part_target_id)->count();
+        $countScore = AppraisalScore::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->count();
         if ($countScore > 0) {
-            AppraisalScore::where('part_target_id', $part_target_id)->delete();
+            AppraisalScore::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->delete();
         }
 
-        $countTransaction = AppraisalTransaction::where('part_target_id', $part_target_id)->count();
+        $countTransaction = AppraisalTransaction::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->count();
         if ($countTransaction > 0) {
-            AppraisalTransaction::where('part_target_id', $part_target_id)->delete();
+            AppraisalTransaction::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->delete();
         }
 
         //save data
@@ -232,6 +249,7 @@ class EvaluateController extends Controller
                             $question->part_target_sub_id = $item->part_target_sub_id;
                             $question->part_index_question_id = $value;
                             $question->part_target_id = $part_target_id;
+                            $question->link_url = $request->input($link_url);
                             $question->created_by = Auth::user()->id;
                             $question->updated_by = Auth::user()->id;
                             $question->save();
@@ -245,6 +263,8 @@ class EvaluateController extends Controller
         $transaction->part_target_id = $part_target_id;
         $transaction->appraisal_transaction_date = date('Y-m-d');
         $transaction->appraisal_transaction_status = '2';
+        $transaction->community_id = session()->get('session_api_community_id');
+        $transaction->community_name = session()->get('community_name');
         $transaction->created_by = Auth::user()->id;
         $transaction->updated_by = Auth::user()->id;
         $transaction->save();
@@ -273,19 +293,19 @@ class EvaluateController extends Controller
         WHERE part_target_id = $part_target_id
         ");
 
-        $countQuestion = AppraisalQuestion::where('part_target_id', $part_target_id)->count();
+        $countQuestion = AppraisalQuestion::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->count();
         if ($countQuestion > 0) {
-            AppraisalQuestion::where('part_target_id', $part_target_id)->delete();
+            AppraisalQuestion::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->delete();
         }
 
-        $countScore = AppraisalScore::where('part_target_id', $part_target_id)->count();
+        $countScore = AppraisalScore::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->count();
         if ($countScore > 0) {
-            AppraisalScore::where('part_target_id', $part_target_id)->delete();
+            AppraisalScore::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->delete();
         }
 
-        $countTransaction = AppraisalTransaction::where('part_target_id', $part_target_id)->count();
+        $countTransaction = AppraisalTransaction::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->count();
         if ($countTransaction > 0) {
-            AppraisalTransaction::where('part_target_id', $part_target_id)->delete();
+            AppraisalTransaction::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->delete();
         }
 
         //save data
@@ -294,7 +314,7 @@ class EvaluateController extends Controller
             $rdo = "rdo_" . $item->part_target_sub_id;
             $comment = "comment_" . $item->part_target_sub_id;
 
-            $score = new AppraisalScore();
+            $score = new AppraisalScore();  
             $score->part_target_sub_id = $item->part_target_sub_id;
             $score->appraisal_score_score = $request->input($rdo);
             $score->appraisal_score_comment = $request->input($comment);
@@ -368,6 +388,8 @@ class EvaluateController extends Controller
         $transaction->part_target_id = $part_target_id;
         $transaction->appraisal_transaction_date = date('Y-m-d');
         $transaction->appraisal_transaction_status = '1';
+        $transaction->community_id = session()->get('session_api_community_id');
+        $transaction->community_name = session()->get('community_name');
         $transaction->created_by = Auth::user()->id;
         $transaction->updated_by = Auth::user()->id;
         $transaction->save();
@@ -382,7 +404,7 @@ class EvaluateController extends Controller
 
     public function show($part_target_id)
     {
-        $part_target = PartTarget::where('part_target_id', $part_target_id)->get();
+        $part_target = PartTarget::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->get();
         
         $part_target_sub = DB::select("
         SELECT 
@@ -398,15 +420,18 @@ class EvaluateController extends Controller
         $part = Part::where('part_id', $part_target[0]->part_id)->get();
         $part_index_score = PartIndexScore::orderBy('part_index_score_order', 'desc')->get();
         $part_index_question = PartIndexQuestion::orderBy('part_index_question_order', 'asc')->get();
+        $part_type_id = $part[0]->part_type_id;
+        $part_type = PartType::where('part_type_id', $part_type_id)->get();
+        $part_type_name = $part_type[0]->part_type_name;
 
         //status
-        $transaction = AppraisalTransaction::select('appraisal_transaction_status')->where('part_target_id', $part_target_id)->get();
+        $transaction = AppraisalTransaction::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->get();
 
         if (!empty($transaction[0]->appraisal_transaction_status)) {
             if ($transaction[0]->appraisal_transaction_status == '2') {
                 //เรียกข้อมูลแบบร่าง
-                $ap_question = AppraisalQuestion::select('part_target_sub_id', 'part_index_question_id')->where('part_target_id', $part_target_id)->get();
-                $ap_score = AppraisalScore::where('part_target_id', $part_target_id)->get();
+                $ap_question = AppraisalQuestion::select('*')->where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->get();
+                $ap_score = AppraisalScore::where(['part_target_id' => $part_target_id, 'created_by' => Auth::user()->id])->get();
 
                 return view('evaluate.show', [
                     'part' => $part,
@@ -417,9 +442,20 @@ class EvaluateController extends Controller
                     'part_target_id' => $part_target_id,
                     'ap_question' => $ap_question,
                     'ap_score' => $ap_score,
+                    'part_type_name' => $part_type_name,
                 ]);
             }
         }
+    }
+
+    public function saveCommunity(Request $request)
+    {
+        $community = $request->evaluate_community;
+        $request->session()->put('evaluate_community', $community);
+
+        return response()->json([
+            'success' => true,
+        ]);
     }
 
 }
